[ refactor ] 네이밍 컨벤션(범용적 단어 대문자로)
