--- api/test.ts
@@ -1,8 +1,0 @@
-import { client } from './common/axios';
-
-export async function test() {
-  const data = await client.get(`/test`, {
-    headers: {},
-  });
-  return data;
-}
