+++ api/wishes/preview/createWishesLink.ts
@@ -1,5 +1,5 @@
-import PATH from '../common/path';
-import { client } from '../common/axios';
+import PATH from '../../common/path';
+import { client } from '../../common/axios';
 import { WishesDataType } from '@/types/wishesDataType';
 
 export async function createWishesLink(wishesData: WishesDataType) {
@@ -8,12 +8,10 @@ export async function createWishesLink(wishesData: WishesDataType) {
   const startDate = wishesData.startDate.replaceAll('-', '.');
   const endDate = wishesData.endDate.replaceAll('-', '.');
 
-  console.log(wishesData);
-
   const data = await client.post(
     `${PATH.API}/${PATH.V1}/${PATH.WISHES}`,
     {
-      imageUrl: wishesData.imageUrl,
+      imageUrl: wishesData.imageURL,
       price: wishesData.price,
       title: wishesData.title,
       hint1: wishesData.hint1,
@@ -32,6 +30,5 @@ export async function createWishesLink(wishesData: WishesDataType) {
       },
     },
   );
-  console.log(data);
   return data;
 }
